+++ charts/ceph-csi-cephfs/templates/provisioner-deployment.yaml
@@ -98,7 +98,6 @@ spec:
             - "--endpoint=$(CSI_ENDPOINT)"
             - "--v=5"
             - "--drivername=$(DRIVER_NAME)"
-            - "--metadatastorage=k8s_configmap"
           env:
             - name: POD_IP
               valueFrom:
@@ -110,10 +109,6 @@ spec:
               valueFrom:
                 fieldRef:
                   fieldPath: spec.nodeName
-            - name: POD_NAMESPACE
-              valueFrom:
-                fieldRef:
-                  fieldPath: metadata.namespace
             - name: CSI_ENDPOINT
               value: "unix:///csi/{{ .Values.provisionerSocketFile }}"
           securityContext:
