+++ deploy/rbd/kubernetes/csi-rbdplugin-provisioner.yaml
@@ -74,10 +74,6 @@ spec:
           env:
             - name: ADDRESS
               value: /csi/csi-provisioner.sock
-            - name: POD_NAMESPACE
-              valueFrom:
-                fieldRef:
-                  fieldPath: metadata.namespace
           imagePullPolicy: "IfNotPresent"
           volumeMounts:
             - name: socket-dir
