+++ docs/deploy-rbd.md
@@ -33,7 +33,6 @@ make image-cephcsi
 | `--nodeid`               | _empty_               | This node's ID                                                                                                                                                                                                                                                                       |
 | `--type`                 | _empty_               | Driver type `[rbd | cephfs]` If the driver type is set to  `rbd` it will act as a `rbd plugin` or if it's set to `cephfs` will act as a `cephfs plugin`                                                                                                                              |
 | `--instanceid`           | "default"             | Unique ID distinguishing this instance of Ceph CSI among other instances, when sharing Ceph clusters across CSI instances for provisioning                                                                                                                                           |
-| `--metadatastorage`      | _empty_               | Points to where legacy (1.0.0 or older plugin versions) metadata about provisioned volumes are kept, as file or in as k8s configmap (`node` or `k8s_configmap` respectively)                                                                                                         |
 | `--pidlimit`             | _0_                   | Configure the PID limit in cgroups. The container runtime can restrict the number of processes/tasks which can cause problems while provisioning (or deleting) a large number of volumes. A value of `-1` configures the limit to the maximum, `0` does not configure limits at all. |
 | `--metricsport`          | `8080`                | TCP port for liveness metrics requests                                                                                                                                                                                                                                               |
 | `--metricspath`          | `"/metrics"`          | Path of prometheus endpoint where metrics will be available                                                                                                                                                                                                                          |
