+++ internal/rbd/driver.go
@@ -75,10 +75,9 @@ func NewIdentityServer(d *csicommon.CSIDriver) *IdentityServer {
 }
 
 // NewControllerServer initialize a controller server for rbd CSI driver
-func NewControllerServer(d *csicommon.CSIDriver, cachePersister util.CachePersister) *ControllerServer {
+func NewControllerServer(d *csicommon.CSIDriver) *ControllerServer {
 	return &ControllerServer{
 		DefaultControllerServer: csicommon.NewDefaultControllerServer(d),
-		MetadataStore:           cachePersister,
 		VolumeLocks:             util.NewVolumeLocks(),
 		SnapshotLocks:           util.NewVolumeLocks(),
 	}
@@ -96,7 +95,7 @@ func NewNodeServer(d *csicommon.CSIDriver, t string, topology map[string]string)
 
 // Run start a non-blocking grpc controller,node and identityserver for
 // rbd CSI driver which can serve multiple parallel requests
-func (r *Driver) Run(conf *util.Config, cachePersister util.CachePersister) {
+func (r *Driver) Run(conf *util.Config) {
 	var err error
 	var topology map[string]string
 
@@ -155,7 +154,7 @@ func (r *Driver) Run(conf *util.Config, cachePersister util.CachePersister) {
 	}
 
 	if conf.IsControllerServer {
-		r.cs = NewControllerServer(r.cd, cachePersister)
+		r.cs = NewControllerServer(r.cd)
 	}
 	if !conf.IsControllerServer && !conf.IsNodeServer {
 		topology, err = util.GetTopologyFromDomainLabels(conf.DomainLabels, conf.NodeID, conf.DriverName)
@@ -166,7 +165,7 @@ func (r *Driver) Run(conf *util.Config, cachePersister util.CachePersister) {
 		if err != nil {
 			klog.Fatalf("failed to start node server, err %v\n", err)
 		}
-		r.cs = NewControllerServer(r.cd, cachePersister)
+		r.cs = NewControllerServer(r.cd)
 	}
 
 	s := csicommon.NewNonBlockingGRPCServer()
