+++ internal/rbd/nodeserver_test.go
@@ -22,25 +22,6 @@ import (
 	"github.com/container-storage-interface/spec/lib/go/csi"
 )
 
-func TestGetLegacyVolumeName(t *testing.T) {
-	tests := []struct {
-		mountPath string
-		volName   string
-	}{
-		{"csi/vol/56a0cc34-a5c9-44ab-ad33-ed53dd2bd5ea/globalmount", "56a0cc34-a5c9-44ab-ad33-ed53dd2bd5ea"},
-		{"csi/vol/9fdb7491-3469-4414-8fe2-ea96be6f7f72/mount", "9fdb7491-3469-4414-8fe2-ea96be6f7f72"},
-		{"csi/vol/82cd91c4-4582-47b3-bb08-a84f8c5716d6", "82cd91c4-4582-47b3-bb08-a84f8c5716d6"},
-	}
-
-	for _, test := range tests {
-		if got, err := getLegacyVolumeName(test.mountPath); err != nil {
-			t.Errorf("getLegacyVolumeName(%s) returned error when it shouldn't: %s", test.mountPath, err.Error())
-		} else if got != test.volName {
-			t.Errorf("getLegacyVolumeName(%s) = %s, want %s", test.mountPath, got, test.volName)
-		}
-	}
-}
-
 func TestGetStagingPath(t *testing.T) {
 	var stagingPath string
 	// test with nodestagevolumerequest
